LayoutTests/fast/dom/HTMLElement/set-false.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<div id="div1" contenteditable="true"></div>
<div id="div2" contenteditable="true">
    <p id="p2"></p>
</div>
<script>
description('When contentEditable is set to "false" string from "true" or inherit, element.contentEditable returns "false" and the element is NOT editable.')

document.getElementById("div1").contentEditable = "false";
document.getElementById("p2").contentEditable = "false";

shouldBe('document.getElementById("div1").getAttribute("contentEditable")','"false"');
shouldBe('document.getElementById("div1").contentEditable', '"false"');
shouldBe('document.getElementById("div1").isContentEditable', 'false');
shouldBe('window.getComputedStyle(div1, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

shouldBe('document.getElementById("p2").contentEditable', '"false"');
shouldBe('document.getElementById("p2").isContentEditable', 'false');
shouldBe('window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify")', '"read-only"');

document.getElementById("div1").style.display= 'none';
document.getElementById("div2").style.display= 'none';
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
