third_party/WebKit/LayoutTests/fast/dom/HTMLElement/set-false-expected.txt

When contentEditable is set to "false" string from "true" or inherit, element.contentEditable returns "false" and the element is NOT editable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("div1").getAttribute("contentEditable") is "false"
PASS document.getElementById("div1").contentEditable is "false"
PASS document.getElementById("div1").isContentEditable is false
PASS window.getComputedStyle(div1, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS document.getElementById("p2").contentEditable is "false"
PASS document.getElementById("p2").isContentEditable is false
PASS window.getComputedStyle(p2, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
